requirements.txt
anthropic>=0.40.0
fastapi>=0.100.0
uvicorn>=0.20.0
chromadb>=0.4.0
sentence-transformers>=2.0.0
pdfplumber>=0.10.0
python-multipart>=0.0.5
pydantic>=2.0.0
python-dotenv>=1.0.0
